requirements.txt
streamlit==1.31.0
watchdog==4.0.0
playwright==1.41.0
beautifulsoup4==4.12.3
torch
transformers==4.37.2
scipy
google-generativeai
python-dotenv==1.0.1
pyyaml==6.0.1
pandas
emoji
underthesea
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
